refactor: Rotating file ordering added to metadata
